refactor[UpdateEtiqueta]
