+++ src/modules/UpdateEtiqueta/dtos/UpdateEtiquetaDTO.ts
@@ -1,0 +1,5 @@
+export interface IUpdateEtiquetaDTO {
+    cookie: string
+    tarefaId: number;
+    etiqueta: string;
+}
